+++ tests/LanguageServer.test.cpp
@@ -1,10 +1,11 @@
 #include "doctest.h"
 #include "LSP/LanguageServer.hpp"
 #include "Protocol/Lifecycle.hpp"
+#include "TestClient.h"
 
 TEST_SUITE_BEGIN("LanguageServer");
 
-LUAU_FASTFLAG(DebugLuauTimeTracing);
+LUAU_FASTFLAG(DebugLuauTimeTracing)
 
 TEST_CASE("language_server_handles_fflags_in_initialization_options")
 {
@@ -58,6 +59,23 @@ TEST_CASE("language_server_lazily_initializes_workspace_folders")
     CHECK(workspaceFolder->isReady);
 
     server.shutdown();
+}
+
+TEST_CASE("language_server_can_process_string_ids")
+{
+    TestClient client;
+    LanguageServer server(&client, std::nullopt);
+
+    auto workspaceUri = Uri::file("project");
+    lsp::InitializeParams initializeParams;
+    std::vector<lsp::WorkspaceFolder> workspaceFolders;
+    workspaceFolders.emplace_back(lsp::WorkspaceFolder{workspaceUri, "project"});
+    initializeParams.workspaceFolders = workspaceFolders;
+
+    server.handleMessage(json_rpc::JsonRpcMessage{"0", "initialize", initializeParams});
+    server.shutdown();
+
+    REQUIRE(client.errorQueue.empty());
 }
 
 TEST_SUITE_END();